+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 * feature: toggling between endianness
+* feature: allow selecting bit-depth of exported integers: 16, 24
 * descriptions for operators (buttons)